+++ docs/Whats-New.md
@@ -73,6 +73,7 @@ New:
 - Implement multiplayer beacons (by ZivDero)
 - Chat improvements (by ZivDero)
 - Port to Syringe (by ZivDero)
+- Add customizable wake animations (by CCHyper, ZivDero)
 - Replace DirectDraw with SDL (by ZivDero, tomsons26, CCHyper)